feat: add auth for task statuses

+++ app/Http/Controllers/TaskStatusController.php
@@ -27,6 +27,8 @@ class TaskStatusController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', TaskStatus::class);
+
         return response()
             ->view('task_status.create');
     }
@@ -39,6 +41,8 @@ class TaskStatusController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', TaskStatus::class);
+
         $data = $this->validate($request, [
             'name' => 'required|unique:task_statuses'
         ]);
@@ -58,6 +62,8 @@ class TaskStatusController extends Controller
      */
     public function edit(TaskStatus $taskStatus)
     {
+        $this->authorize('update', $taskStatus);
+
         return response()
             ->view('task_status.edit', compact('taskStatus'));
     }
@@ -71,6 +77,8 @@ class TaskStatusController extends Controller
      */
     public function update(Request $request, TaskStatus $taskStatus)
     {
+        $this->authorize('update', $taskStatus);
+
         $data = $this->validate($request, [
             'name' => 'required|unique:task_statuses'
         ]);
@@ -89,6 +97,8 @@ class TaskStatusController extends Controller
      */
     public function destroy(TaskStatus $taskStatus)
     {
+        $this->authorize('delete', $taskStatus);
+
         $taskStatus->delete();
 
         return redirect()
